+++ src/run_sp.h
@@ -27,12 +27,12 @@ std::shared_ptr <HeapDesc> getHeapImpl(const std::string& heap_type);
 
 size_t make_vert_map (const Rcpp::DataFrame &vert_map_in,
         const std::vector <std::string> &vert_map_id,
-        const std::vector <unsigned int> &vert_map_n,
-        std::map <std::string, unsigned int> &vert_map);
+        const std::vector <size_t> &vert_map_n,
+        std::map <std::string, size_t> &vert_map);
 
 void make_vert_to_edge_maps (const std::vector <std::string> &from,
         const std::vector <std::string> &to, const std::vector <double> &wt,
-        std::unordered_map <std::string, unsigned int> &verts_to_edge_map,
+        std::unordered_map <std::string, size_t> &verts_to_edge_map,
         std::unordered_map <std::string, double> &verts_to_dist_map);
 
 size_t get_chunk_size (const size_t nfrom);
